Refactor CI tasks
