+++ easy_infra/__init__.py
@@ -1,0 +1,6 @@
+"""
+init
+"""
+__maintainer__ = "Seiso"
+__copyright__ = "(c) 2020 Seiso, LLC"
+__project_name__ = "easy_infra"
